nlp-engine/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pymongo>=4.0.0
nltk>=3.8.1
scikit-learn>=1.2.0
numpy>=1.24.0
pandas>=1.5.0
joblib>=1.2.0
slowapi>=0.1.8
pytest>=7.0.0
httpx>=0.24.0
